Remove build-with-cachix CI step
This step was practically only testing the previous release, not the
current build. The existing test-cachix step already does a similar test
for the current build.

One thing this step did that test-cachix didn't was actually pull
something from cachix. Though, we can easily add this to test-cachix
instead if desired.